# docs/api-contract.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Architect Canonical API Contract",
    "version": "1.0.0",
    "x-artifact-type": "CANONICAL_API_CONTRACT",
    "x-owner": "architect"
  },
  "paths": {
...
    "/api/v1/profiles/{username}": {
      "get": {
        "operationId": "Profile_get_api_v1_profiles_username",
        "summary": "Profile",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "x-error-semantics": [
          {
            "status": 400,
            "semantic": "validation_error"
          },
          {
            "status": 401,
            "semantic": "authentication_required"
          },
          {
            "status": 403,
            "semantic": "forbidden"
          },
          {
            "status": 404,
            "semantic": "not_found_or_hidden"
          },
          {
            "status": 409,
            "semantic": "conflict"
          },
          {
            "status": 500,
            "semantic": "server_error"
          }
        ],
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
    "/api/v1/profiles/{username}/followers": {
      "get": {
        "operationId": "SocialGraph_get_api_v1_profiles_username_followers",
        "summary": "SocialGraph",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "x-error-semantics": [
          {
            "status": 400,
            "semantic": "validation_error"
          },
          {
            "status": 401,
            "semantic": "authentication_required"
          },
          {
            "status": 403,
            "semantic": "forbidden"
          },
          {
            "status": 404,
            "semantic": "not_found_or_hidden"
          },
          {
            "status": 409,
            "semantic": "conflict"
          },
          {
            "status": 500,
            "semantic": "server_error"
          }
        ],
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/profiles/{username}/following": {
      "get": {
        "operationId": "SocialGraph_get_api_v1_profiles_username_following",
        "summary": "SocialGraph",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "x-error-semantics": [
          {
            "status": 400,
            "semantic": "validation_error"
          },
          {
            "status": 401,
            "semantic": "authentication_required"
          },
          {
            "status": 403,
            "semantic": "forbidden"
          },
          {
            "status": 404,
            "semantic": "not_found_or_hidden"
          },
          {
            "status": 409,
            "semantic": "conflict"
          },
          {
            "status": 500,
            "semantic": "server_error"
          }
        ],
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/profiles/{username}/follow": {
      "post": {
        "operationId": "Follow_post_api_v1_profiles_username_follow",
        "summary": "Follow",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "x-error-semantics": [
          {
            "status": 400,
            "semantic": "validation_error"
          },
          {
            "status": 401,
            "semantic": "authentication_required"
          },
          {
            "status": 403,
            "semantic": "forbidden"
          },
          {
            "status": 404,
            "semantic": "not_found_or_hidden"
          },
          {
            "status": 409,
            "semantic": "conflict"
          },
          {
            "status": 500,
            "semantic": "server_error"
          }
        ],
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "Follow_delete_api_v1_profiles_username_follow",
        "summary": "Follow",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "x-error-semantics": [
          {
            "status": 400,
            "semantic": "validation_error"
          },
          {
            "status": 401,
            "semantic": "authentication_required"
          },
          {
            "status": 403,
            "semantic": "forbidden"
          },
          {
            "status": 404,
            "semantic": "not_found_or_hidden"
          },
          {
            "status": 409,
            "semantic": "conflict"
          },
          {
            "status": 500,
            "semantic": "server_error"
          }
        ],
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}